JavaScript: 1. Created separate js files to handle main.html and login.html page loads. 2. Added a test for localStorage, storing the value "Bob" into the Username Item.
HTML: Added IDs to HTML code, but did not change any core characteristics.

PHP: No functionality changes yet, just added some comments, etc. This file will change because JavaScript is sending data using the POST and JSON functionality/formatting. So the PHP script must be able to interpret it properly. But that is on the TODO list.